Added Information
While downloading URLs.IMDB (which is a huge data-set), I didn't knew that it was being extracted after being downloaded, because no such information was printed to console of file being extracted. And as I was using Google Colab (which I think most of the students prefer), it takes a lot of time to extract the file contents. And I cancelled the process 4 times before acknowledging what's going on.

I think it's better that we just print the the on-going process for everyone's benefit.
(I added a flag just not to surprise anyone, though I don't think it's necessary)